# TRY perf(core): avoid borrows / multiple `get_slot` calls in poll_event_loop