Async ops benchmark:

`main`:
```
op_void_async:       	n = 1000000, dt = 0.279s, r = 3584229/s, t = 279ns/op
read_128k:           	n = 50000, dt = 7.833s, r = 6383/s, t = 156660ns/op
```

This PR:
```
op_void_async:       	n = 1000000, dt = 0.258s, r = 3875969/s, t = 258ns/o
read_128k:           	n = 50000, dt = 7.164s, r = 6979/s, t = 143280ns/op
```

## note

opening pr for visibility, not super happy with improvements. Need to try a few more things.